requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.2.0
frozenlist==1.5.0
idna==3.10
lxml==5.3.0
multidict==6.1.0
pika==1.3.2
propcache==0.2.1
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.18.3
